SqlPractice
I created a database and used it to create a table. I inserted values into the table and displayed all the records, as well as some specific records. I added a new column to the table, updated it with values, and deleted an unnecessary column and a record. I also changed the data type of a column.